Rename the Quartz.Serialization.Json namespaces to Quartz.Serialization.SystemTextJson
These are the System.Text.Json types, which merged into the core package in 4.0.
The namespace they sat in was named after the 3.x Newtonsoft package, which is
retired — so the core package held a Json namespace that was not the Json
package's, while the Newtonsoft package next to it is spelled out in full.
Quartz.JsonConfigurationExtensions becomes SystemTextJsonConfigurationExtensions
for the same reason.

Nothing persists a serializer's type name and the STJ discriminators are logical
short names ("CronTrigger", "AnnualCalendar"), so stored data is unaffected and
no SimpleTypeLoadHelper entry is needed.

The migration guide carries a warning the rename makes necessary: 3.x's
*Newtonsoft* package owned Quartz.Serialization.Json.Triggers, so porting one of
its serializers by editing the using directive alone now binds to the
System.Text.Json base class and fails on the override signatures.
