+++ src/Quartz.Dashboard/Components/Shared/DisplayValueHelper.cs
@@ -21,7 +21,7 @@ using System.Globalization;
 using System.Reflection;
 using System.Text.Json;
 
-using Quartz.Serialization.Json;
+using Quartz.Serialization.SystemTextJson;
 
 namespace Quartz.Dashboard.Components.Shared;
