+++ src/Quartz.Tests.Integration/RAMJobStoreTest.cs
@@ -1,8 +1,8 @@
 using System.Text.Json;
 
 using Quartz.Impl.Triggers;
-using Quartz.Serialization.Json;
-using Quartz.Serialization.Json.Triggers;
+using Quartz.Serialization.SystemTextJson;
+using Quartz.Serialization.SystemTextJson.Triggers;
 using Quartz.Extensibility;
 
 namespace Quartz.Tests.Integration;
