+++ src/Quartz/Configuration/SchedulerScopedServiceProvider.cs
@@ -7,7 +7,7 @@ using Quartz.Core;
 using Quartz.Impl;
 using Quartz.Impl.AdoJobStore;
 using Quartz.Impl.AdoJobStore.Common;
-using Quartz.Serialization.Json;
+using Quartz.Serialization.SystemTextJson;
 using Quartz.Extensibility;
 
 namespace Quartz.Configuration;
