+++ src/Quartz/Serialization/SystemTextJson/Converters/CronExpressionConverter.cs
@@ -3,7 +3,7 @@ using System.Text.Json.Serialization;
 
 using Quartz.Util;
 
-namespace Quartz.Serialization.Json.Converters;
+namespace Quartz.Serialization.SystemTextJson.Converters;
 
 internal sealed class CronExpressionConverter : JsonConverter<CronExpression>
 {
