+++ src/Quartz/Serialization/SystemTextJson/Triggers/RecurrenceTriggerSerializer.cs
@@ -2,7 +2,7 @@ using System.Text.Json;
 
 using Quartz.Impl.Triggers;
 
-namespace Quartz.Serialization.Json.Triggers;
+namespace Quartz.Serialization.SystemTextJson.Triggers;
 
 public sealed class RecurrenceTriggerSerializer : TriggerSerializer<RecurrenceTriggerImpl>
 {
